Add missing semicolons
When using the non-minified build concatenated with other libs in a
single file, additional libraries immediately after Lunr will throw
error if wrapped in IIFE.

For example, if we have the following:

	A.prototype.hello = function(){}
	(function(){})();

it will throw:

	TypeError: undefined is not a function 

Since it'll interpret `(function(){})` as calling the function assigned
to `A.prototype.hello`.

This commit adds semicolons to all functions -regardless they're capable
of reproducing the error above- to maintain uniformity.