Append to the end of the list where possible.
Vectors are implemented as a linked-list, almost all of the time
dimensions will be inserted in ascending order. By keeping track of the
end of the list we can quickly append dimensions, whilst still being
able to handle dimensions added out of order.

This leads to a slight increase in performance when scoring documents as
part of a search.