Stop storing document ref twice in token store.
The token store has a docs object that is used as a key value pair for
looking up the term frequency associated with the given term and
document. Before this change the reference to the document was stored
twice, once as the key of the `docs` object in the token store, and once
as the `ref` property of each entry in the docs object.

Removing this duplicate resulted in a ~20% reduction in the serialized
index used for the examples on the lunr site.