Spike storing tf values as rational numbers.
This commit is an experiment in storing the values generated for tf as
rational numbers, e.g. instead of storing a tf value of 0.5 we would
store 1/2.

The motiviation of this experiment was to reduce the size of the
serialized index, since most tf values have many decimal places the idea
was that by serialising this as a more succinct rational would save
space in the serialised index.

The assumption proved true, using the rational representation reduced
the index size of the example index by ~20% before gzip, it was less
impressive after compression.

The increase in load time made the current implementation of this
approach unworkable. Load times increased by nearly two orders of
magnitude. On the example index the load time increased from ~2ms to
around 100ms. Also search performance was also negatively impacted.

Load time increase is expected as the whole tree must be traversed to
convert the rational back to an instance of Fraction. I'm not sure what
the decrease in search performance was from.