Cut recursion in SortedSet#indexOf & locationFor.
Reduce the overhead of recursive function calls from SortedSet#indexOf
and SortedSet#locationFor. On their own these changes make a
considerable performance gain, this improvement is more modest in the
context of the public Index#search and Index#add interface but is still
noticeable.

As well as the speed performance this change should also prevent these
methods blowing the stack on very large SortedSets.