Move empty token filter from trimmer to pipeline.
The pipeline should ensure that the index gets clean data to work with,
in this case clean data means that no token should be empty. The
pipeline already included logic to filter out tokens that are undefined.
The pipeline function lunr.trimmer contained logic to exclude empty
strings.

The implementation of trimmer is not generic enough to be used with all
languages, specifically non latin character languages are not well
handled by the current implementation, and therefore in some cases users
choose to omit the trimmer from their pipeline. This can lead to issues
where empty tokens are introduced into the index.

This change moves the logic for filtering out empty tokens from the
trimmer to the pipeline itself, in the same way that the pipeline
removes undefined tokens, it now also removes empty tokens. With this
change the pipeline will guarantee that the index sees clean data,
regardless of whether the trimmer is part of the pipeline, it also
allows the trimmer function to be more focused on just trimming
whitespace.