refactored all add() calls from SortedSet to only pass single values
before this fix, the index passed all tokens via an apply call, pushing all parameters on the stack.
This behaviour causes a RangeError when more tokens are added than the JS implementation can handle (Firefox 44 -> 262143, Node.js -> 65535, evaluated using this test: http://stackoverflow.com/questions/22747068/is-there-a-max-number-of-arguments-javascript-functions-can-accept)
this fix changes the code so that now only a single value gets passed per call. This way no change to the current API is neccesary.

the code was refactored in a way that it now passes only a single argument at a time, avoiding the usage of apply

The only parts of the code that still use apply are the EventEmitters and the Plugin subsystem which should be non critical for this type of error.