Make tokenizer a property of an index.
Before this change the tokenizer was a global property of lunr, every
index would use the same tokenizer. It was possible to change the
tokenizer but not on a per index basis.

This change gives each instance of lunr.Index its own tokenzier,
defaulting to the existing lunr.tokenizer implementation. Each index can
change the tokenizer it uses without affecting any other index instance.

To ensure that indexes with custom tokenizers can be serialied and
de-serialised custom tokenizers can be regsitered in much the same way
as a custom text processing function is registered in lunr.Pipeline.