Improve the performance of Index.add.
The loops in Index.add can easily be called hundreds of thousands of times for
large documents. Larger documents can perform particularly badly[1] under
Chrome. Removing the usage of reduce and filter in favour of inline native for
loops can lead to a 20-30% performance increase in Chrome. Firefox also sees a
similar but much smaller impovement (~5%).

[1]: https://github.com/mkdocs/mkdocs/issues/859#issuecomment-192187096