Refactor Vector implementation.
Instead of using a linked list we use a flat array where the odd
elements are the indexes and the even elements are the values.

This roughly halves the amount of space required for the vector,
sacrificing a very small amount of performance. The _very_ small
performance drop during calculating the dot product surprised me, I
would have thought it would have been quicker, calculating the magniture
is about twice as quick though with this implementation.