Allow TokenSet#toString to be optimised by V8.
The token set object contains an object `edges` that is being used as a
hash. This prevents V8 from optimising the function [1]. Knowing that
the `edges` object is always 'hash like' we get all its keys using
Object.keys and iterate this list using a regular for loop.

Some quick benchmarks showed a small increase in performance from this,
but the function is now being optimised, which means that we should be
able to get 'free' performance improvements from V8, and possibly other
JavaScript engines, in the future.

[1]
http://stackoverflow.com/questions/24414809/iterating-through-javascript-hashtable-without-causing-optimization-loss