Add search pipeline.
A default search pipeline contains just a stemmer. When searching terms
are passed through the search pipeline and the resulting strings are
used for lookups in the token set.

Some terms will not be passed through the pipeline, primarily those
which contain a wildcard, but any term can bypass the pipeline by
setting the relevant option on the query term.