Dry IDF calculation.
Centralise the calculation of inverse document frequency, a single
function is shared between the builder and the index.

This also stops storing the IDF in the inverted index. This stored value
was never actually being used. By removing it we get a slight speed up
in the builder as well as a noticeable speed up in searching. A small
amount of space will also be saved when serialising and keeping the
index in memory.