Correct spelling of separator.
So, I cannot spell :(

Sadly this mispelling was in the public API for customising the
separator for the tokenizer. Backwards compatibility is retained by
keeping the misspelt property, but preferring the correctly spelt
separator property.

Integrations that make use of the old property will still work, but this
property is now deprecated and will be removed in the next major version
release. Customising the tokenizer should be done by setting the new
property - lunr.tokenizer.separator.

On the plus side maybe this will help me remember to spell separator
correctly from now on!