Upgrade PhantomJS.
Yes, the previous version was probably very old, but it worked. A fresh
install of the dependencies failed trying to install phantomjs. I don't
know how this happens, but it's pretty borked and I don't have time to
fix this now.

Instead upgrading to the next minor version resolves the problem.