Change structure of MatchData metadata.
Before it was not possible to know which term some metadata was
associated to. With this change the metadata is collection of metadata
addressable by term and field, e.g. metadata[term][field][key].

Specifically without this it made highlighting matches difficult as the
positions were not in order, after this change matches are grouped by
term and field and therefore are always in order.

This might make it slightly more hardwork for consumers to use the data
but as common access patters arise convinience methods can be added.