Remove empty message detail from QueryParseError.
When there are problems parsing a boost or edit distance from a query
the lexeme found is a zero length string. The previous error messages
for these errors tried to indicate what was found by the parser, a zero
length string is not useful in these cases.

Arguably the parsing/lexing process could provide what was found instead
of the expected numbers, this is a larger change though.