Ensure that query terms are lower case.
The index stores all tokens as lower case, this is handled by the
tokeniser. Queries are not passed through the tokeniser though, instead
they are parsed by the QueryParser.

In this change the QueryParser is modified to down case all query terms,
this leaves any up case fields untouched.

Before this change a search term that in upper case would result in no
matches, even if a document contained that term. This is especially
confusing if the original document had the term in upper case.