Fix bug that allowed extra entries in token sets.
TokenSet#toString is used to identify candidates during minimisation.
Previously this was a recursive function that incorporated the finality
of a TokenSet, its edges and the toString of all its children.

This lead to a bug where TokenSets were minimised incorrectly, leading
to extraneous entries in the TokenSet. When these were looked up in the
inverted index, there was no corresponding entry, leading to no method
errors.

By giving each token set an auto incrementing id we can be sure when we
are referring to the same TokenSet during minimsation.

This issue was detected in https://github.com/olivernn/lunr.js/issues/68