Use objects without prototypes in lieu of HashMap
Lunr makes heavy use of objects in place of a native HashMap. In most
cases this is fine, but can cause problems when the keys to be stored
clash with properties in Object.prototype.

This change swaps the instances where we check for the existance of a
key within an object to use an object without prototype, created using
Object.create.

This has some implications for support in old browsers, they would need
to use a library that provides a polyfil for this behaviour of
Object.create. Most browsers now seem to support this so I think it is a
trade off worth making.