Make separator used in lexing match tokeniser
Previously only a space " " was considered a term separator when lexing
and parsing a search string. This did not match the tokenisers idea of
what was considered a token separator. This lead to issues where it was
not possible to search for some terms within an index. Specifically a
document containing a hyphenated token, e.g. "foo-bar" could not be
searched for as the query parser would convert this to two separate
terms "foo" and "bar" where as in the index this would be a single term
"foo-bar".

This commit adds a property on the lunr.QueryLexer for the correct
separator to use, it currently points to lunr.tokenizer.separator so
that if a user changes the global separator for tokenising it will be
reflected when parsing queries. It may be required to make this more
configurable in the future, but for now this should suffice.