Reduce term score precision in vectors
This change reduces the number of decimal places used to represent the
score of a term within a document vector. Through testing this has been
seen to not have any impact to the relevance of the search results but
does lead to a ~15% size reduction when serialising indexes, before and
after compression with gzip.