Add lunr.Vector#upsert method
Before lunr.Vector only supported insert, and this method would fail if
trying to insert a value with a duplicate index. When building the
document vectors this is the correct behaviour, however, for query
vectors there may be duplicate query terms in the query. Instead of
failing, the values should be combined.

In the future there should probably be some kind of query optimisation
that runs to execute queries in a smarter way, e.g. not querying for the
same term twice, but this is more complex than I want to tackle right
now.

Instead this change unblocks some current usecases for lunr and doesn't
close any doors to further query optimisations.