Add support for adding wildcards to clause term
Introduces a lunr.Query.wildcard property with flags for defining
whether to add leading or trailing wildcards to a clause term. Flags can
be combined to add leading and trailing wildcards.

The term is checked for existing wildcards to prevent double wildcards
and by default the wildcard property is set to none.