Create a query vector per field for scoring
Previously a single query vector was created for use in scoring
documents against a query. This could lead to some incorrect results
when performing queries with boost specific to certain fields.

If a query term has different boost for different fields the approach of
using a single vector will mean that the same boost is used for both
fields.

Instead we create a vector per field, this prevents a boost from being
applied to terms in different fields.