Prevent duplicate inverted index lookups
If a query contains duplicate terms for a given field we only need to
retrieve the matching documents once. Before this change a duplicate
query term field combination would be looked up in the inverted index
multiple times.

This is especially a problem when constructing a query as recomended for
perfoming type ahead search. The combination of exact match, prefix
match and fuzzy match is able to produce multiple duplicate query terms.

This change keeps a cache of already seen term field combinations and
uses this cache to only lookup matching documents for that term field
combination once.