Cache results of idf calculation during build
Generating vectors is the most time consuming aspect of building an
index, specifically the idf calculation. Before this change the inverse
document frequency of a term would be calculated more than once if that
term appeared in multiple documents/fields. This was not cached as it
was beleived that the cost of calculating the idf was cheap, however new
data shows that an improvement in build performance would be possible by
caching idf values.

https://github.com/olivernn/lunr.js/issues/305