Stop calculating similarity score unnecessarily
In cases where we know a document/field is not going to make it into the
results due to a terms presence (either required or prohibited) there is
no point in calculating a similarity score.

This leads to a minor performance improvement for queries that contain
either required or prohibited terms.