Add field metadata to tokens
When building and searching, tokens passing through the relevant
pipeline will have an additional entry in the metadata, fields.

At build time this will be the field in which the token was found, at
search time this will be the fields this term is going to be searched
in.

This additional entry in the token metadata will allow users to
customise their pipeline, performing different processing depending on
the field a specific token is from.