Fix bug when using field scoped required terms
The correct behaviour when searching for a required term without field
scoping is that the term must appear in ANY of the fields. However the
previous implementation meant that a term must appear in ALL of the
fields.

Additionally there was a bug in how the required field matches were
combined which lead to more results being returned than should have.