Modify lunr.Vector#similarity calculation
Removing otherVector.magnitude from the divisor allows similarity to
better accomodate boosted term scores.

This allows existing query-time boosts to be more effective as well as
paving the way for re-introducing build-time boosts.

The actual values returned when scoring will have changed as a result of
this modification. However the relative ranking of search results before
and after this change should not change. The score value in the search
results is an implementation detail used purely for sorting.