Fix issue #279 (bug with Safari) (#361)
* Fix issue #279 (bug with Safari)

Calling any method defined on String.prototype on the expanded term
seems to force the string to be properly represented, fixing an issue
affecting Safari users.

See https://github.com/olivernn/lunr.js/issues/279

* fix issue #279 at the source, on TokenSet.prototype.toArray

it turns out that a specific string concatenation in
TokenSet.prototype.toArray sometimes results in a corrupted string in
Safari. It is fixed by calling any String.prototype method on it, which
forces the string to the correct representation.

The previous commit did the same, but this commit moves the fix closer
to the source of the problem. It could be applied exactly at the point
of the problematic concatenation, but that would result in some
unnecessary (if inexpensive) calls, so it is instead when pushing each
result string in the returned array.

* remove changes to generated distribution file