Fix fuzzy string token set construction
A couple (more) bugs in how fuzzy string or edit distance based token
sets are constructed. This time there was issues with how fuzzyness at
the end of a string is matched.

A couple of fixes were applied; the change that was previously made to
fix this was applied to _all_ relevant parts of the function.
Additionally the way delete edits were addded to the token set was
slightly simplified. Finaly some redundant conditionals were removed to
hopefully simplify the logic.